+++ CHANGELOG.md
@@ -5,3 +5,4 @@ Require Create: Dragons Plus 1.5.0
 #### Update
 * Now you can throw item to Mechanical Grindstone to process it
 * Improve English grammar. clarity, and consistency (by @Zain)
+* Added a few more explanation on ponder